+++ src/Chunkyard/Options/KeepOptions.cs
@@ -8,16 +8,16 @@ namespace Chunkyard.Options
     {
         public KeepOptions(
             string repository,
-            IEnumerable<int> logPositions)
+            int latestCount)
         {
             Repository = repository;
-            LogPositions = new List<int>(logPositions);
+            LatestCount = latestCount;
         }
 
         [Option('r', "repository", Required = true, HelpText = "The repository")]
         public string Repository { get; }
 
-        [Option('l', "log-positions", Required = true, HelpText = "The log positions to keep")]
-        public IEnumerable<int> LogPositions { get; }
+        [Option('l', "latest", Required = true, HelpText = "The count of latest log positions to keep")]
+        public int LatestCount { get; }
     }
 }